soteria-backend-main/pyproject.toml
[tool.poetry]
name = "soteria-backend"
version = "0.1.0"
description = ""
authors = ["DataPeace <info@datapeace.in>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
django-configurations = "^2.3.1"
django-environ = "^0.8.1"
djangorestframework = "3.13.1"
django-cors-headers = "3.11.0"
django-storages = "1.12.3"
django-filter = "21.1"
celery = "5.2.7"
django-model-utils = "4.2.0"
gunicorn = "^20.1.0"
gevent = "21.12.0"
requests = "^2.28.0"
sentry-sdk = "^1.6.0"
Pillow = "^9.1.0"
PyJWT = "^2.4.0"
drf-exceptions-hog = "^0.2.0"
djangorestframework-simplejwt = "5.0.0"
django-ipware = "^4.0.2"
django-tenants = "^3.4.2"
psycopg2 = "^2.9.3"
django-extensions = "3.1.5"
redis = "4.3.1"
django-redis = "5.2.0"
qrcode = "^7.3.1"
cryptography = "2.8"
boto3 = "^1.24.84"
django-timezone-field = "^5.0"
pandas = "^1.5.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
ipython = "^8.2.0"
django-debug-toolbar = "^3.4.0"

[tool.black]
# File filtering is taken care of in pre-commit.
line-length=100
target-version=["py38"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"